+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Additional Calypso requirements related to Stored Value operations (issue [#59]).
 - Additional Calypso requirements related to abortion of a secure session (issue [#57]).
+- Additional Calypso requirement related to the analysis of the APDU response length (issue [#62]).
 
 ## [2.1.0] - 2022-02-01
 ### Added
@@ -60,6 +61,7 @@ It also brings many major API changes.
 [2.0.1]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/eclipse/keyple-card-calypso-java-lib/releases/tag/2.0.0
 
+[#62]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/62
 [#59]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/59
 [#57]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/57
 [#55]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/55